requirements.txt
pandas == 2.1.1  
dvc == 3.22.0
notebook 
numpy == 1.26.0
matplotlib == 3.8.0 
seaborn == 0.12.2 
python-box==6.0.2
pyYAML == 6.0.1
tqdm == 4.66.1
ensure==1.0.2
joblib == 1.3.2
types-PyYAML == 6.0.12.11 
pymysql == 1.1.0 
##scipy
Flask == 2.3.3
Flask-Cors == 4.0.0
Pillow == 10.1.0 
tensorflow == 2.10.0